Do WL roundtrip after destroying surface & before moving cursor
This makes sure that the overlay surface is destroyed before we attempt
to move the cursor. If we don't do that, on Hyprland, we move the cursor
but then when the focus is gained, the cursor is moved back to the
center of the newly focused window. With this change, the focus is
gained, and only after that the cursor moves to the right position.